Now getting libgl from conda-forge instead of system package.
The documentation on core dependency trees that was linked to
in the comment now says "packages dependent on OpenGL and/or
libGL should no longer use CDTs. Instead, use the host dependency
libgl-devel." I verified that tile generation (which is what we
use opencv for) still works in this container.